Refactor code into leaner main file
